Fix Respond status code for gzip
