+++ internal/app/app.go
@@ -1,11 +1,8 @@
 package app
 
 import (
-	"compress/gzip"
-	"io"
 	"log"
 	"net/http"
-	"strings"
 
 	"github.com/e-pas/yandex-praktikum-shortener/internal/app/config"
 	"github.com/e-pas/yandex-praktikum-shortener/internal/app/endpoint"
@@ -34,7 +31,7 @@ func New() (*App, error) {
 	a.r.Use(middleware.RealIP)
 	a.r.Use(middleware.Logger)
 	a.r.Use(middleware.Recoverer)
-	a.r.Use(gzipHandle)
+	a.r.Use(config.GzipHandle)
 
 	a.r.Get("/{id}", a.e.Get)
 	a.r.Get("/info", a.e.Info)
@@ -48,31 +45,4 @@ func (a *App) Run() error {
 	log.Println("service running")
 	err := http.ListenAndServe(a.c.Listen, a.r)
 	return err
-}
-
-type gzipWrite struct {
-	http.ResponseWriter
-	gzWriter io.Writer
-}
-
-func (gz gzipWrite) Write(buf []byte) (int, error) {
-	return gz.gzWriter.Write(buf)
-}
-
-func gzipHandle(next http.Handler) http.Handler {
-	fn := func(w http.ResponseWriter, r *http.Request) {
-		if !strings.Contains(r.Header.Get("Accept-Encoding"), "gzip") {
-			next.ServeHTTP(w, r)
-			return
-		}
-		gzw := gzip.NewWriter(w)
-		defer gzw.Close()
-		gz := gzipWrite{
-			ResponseWriter: w,
-			gzWriter:       gzw,
-		}
-		w.Header().Set("Content-Encoding", "gzip")
-		next.ServeHTTP(gz, r)
-	}
-	return http.HandlerFunc(fn)
 }
